cast pthread to ulong in cpu_seqid
This patch adds a cast of pthread_self in CPU_SEQID definition.
This fixes a build time error, resulting from comparing types
pthread_t and int, on alternate platforms.

+++ include/sys/zfs_context.h
@@ -614,7 +614,7 @@ extern void delay(clock_t ticks);
 #define	minclsyspri	60
 #define	maxclsyspri	99
 
-#define	CPU_SEQID	(pthread_self() & (max_ncpus - 1))
+#define	CPU_SEQID	((unsigned long)pthread_self() & (max_ncpus - 1))
 
 #define	kcred		NULL
 #define	CRED()		NULL
